feat(core,storage,index,bolt): widen Date to i64 days — full openCypher year range, TCK 100% (#141)
Date stored i32 days (~+-5.88M-year span); openCypher requires +-999,999,999
years (Temporal10 [9][10]). Widened to i64 days end-to-end with a crash-safe,
backward-compatible on-disk format change.

- core/temporal_calc.rs: Date.days_since_epoch i32 -> i64; MAX_ABS_YEAR
  999_999_999; date-only localdatetime(...) defaults to midnight (CIP, needed by
  [10]).
- storage/valenc.rs: NEW TAG_DATE_WIDE (13) with an 8-byte i64 LE body; encode
  always emits it; decode reads the wide tag AND still reads legacy TAG_DATE (6)
  as 4-byte i32 sign-extended. Self-describing width => NO in-place migration,
  no byte already on disk changes meaning (crash-safe). Same for Date in lists.
- index/keycodec.rs: Date key i32 -> i64 offset-binary, order-preserving;
  secondary indexes are in-memory and rebuilt at open, so no legacy on-disk key
  risk.
- bolt/packstream.rs: removed the i32 narrowing on decode (PackStream INTEGER is
  i64).
- rest/cli date formatting widened; new tests incl. legacy_narrow_date_image_
  still_decodes (backward-compat) and full-range round-trips.
- tck.rs: baseline 3912 -> 3914.

Temporal10 [9][10] pass => TCK 3914/3914 (100.00%). Cluster-1 durability
RE-CERTIFIED after the format change: cargo test --workspace --release 2378
passed/0 failed; loom 4/4; crash_recovery + aries_recovery green; legacy-date
backward-compat proven. clippy --workspace -D warnings clean.

+++ crates/graphus-bolt/tests/packstream_roundtrip.rs
@@ -65,7 +65,7 @@ fn leaf_value() -> impl Strategy<Value = Value> {
 /// Every temporal `Value` class.
 fn temporal_value() -> impl Strategy<Value = Value> {
     prop_oneof![
-        any::<i32>().prop_map(|d| Value::Date(Date {
+        any::<i64>().prop_map(|d| Value::Date(Date {
             days_since_epoch: d
         })),
         // nanos-of-day stays within a day.
